This is a bit better, but there are still some “bounce” issues
